+++ scripts/static-build.sh
@@ -9,10 +9,10 @@ cd $(dirname "$0")/..
 
 set -o pipefail
 
+# Use `docker-alpine-image` field to replace ocamlpro/ocaml:4.13
+# and `docker-alpine-packages` to add more apk packages
 git ls-files -z | xargs -0 tar c | \
 docker run --rm -i \
-    # Use `docker-alpine-image` field to replace ocamlpro/ocaml:4.13
-    # and `docker-alpine-packages` to add more apk packages
     ocamlpro/ocaml:4.13 \
     sh -uexc \
       'tar x >&2 &&