+++ CHANGES.md
@@ -1,4 +1,9 @@
 
+## v0.2.0
+
+* Add option `--skip-dir DIR` to skip a directory
+* Allow comma separated list of header ids in `--replace HEADER`
+
 ## v0.1.0 ( 2023-04-16 )
 
 * Initial commit